requirements.txt
discord.py>=2.3.0
gspread>=5.10.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
aiohttp>=3.8.0